refactoring -> created AbstractEnemy
